Account_Managemnet_Projects/Project1/requirements.txt
google-adk
vertexai
python-dotenv
streamlit
requests
fastapi
uvicorn
google-generativeai
psycopg2
bcrypt